move code mode & update root state
