+++ editor/components/editor/editor-hierarchy/index.ts
@@ -1,0 +1,1 @@
+export { EditorLayerHierarchy } from "./editor-hierarchy";
